Move to umbrella structure in preparation for separate application for data fetching
