+++ .gitignore
@@ -8,12 +8,12 @@
 erl_crash.dump
 
 # Static artifacts
-/node_modules
+/apps/action/node_modules
 
 # Since we are building assets from web/static,
 # we ignore priv/static. You may want to comment
 # this depending on your deployment strategy.
-/priv/static/
+/apps/action/priv/static/
 
 # The config/prod.secret.exs file by default contains sensitive
 # data and you should not commit it into version control.
@@ -21,8 +21,8 @@ erl_crash.dump
 # Alternatively, you may comment the line below and commit the
 # secrets file as long as you replace its contents by environment
 # variables.
-/config/prod.secret.exs
-/config/dev.secret.exs
+prod.secret.exs
+dev.secret.exs
 
 npm-debug.log
