+++ apps/action/priv/gettext/errors.pot
