+++ apps/action/web/static/js/app.js
